refactor: move filtering logic to its own file
